+++ resolwe/process/tests/processes/python_json.py
@@ -8,13 +8,6 @@ class PythonProcessJson(Process):
     name = "Python Process that uses JSON field"
     process_type = "data:python"
     version = "0.1.2"
-    requirements = {
-        "executor": {
-            "docker": {
-                "image": "resolwe/base:ubuntu-20.04",
-            }
-        }
-    }
 
     class Input:
         """Input fields."""