+++ resolwe/process/tests/test_python_process.py
@@ -62,7 +62,7 @@ class PythonProcessTest(ProcessTestCase):
                 "expression-engine": "jinja",
                 "executor": {
                     "docker": {
-                        "image": "resolwe/base:ubuntu-20.04",
+                        "image": "resolwe/base:ubuntu-20.04dev",
                     }
                 },
             },
@@ -162,7 +162,7 @@ class PythonProcessTest(ProcessTestCase):
         self.assertEqual(data.output["dir_output"], {"dir": "test/"})
         self.assertEqual(data.output["input_data_name"], "bar")
         self.assertEqual(data.output["input_entity_name"], "Data with entity")
-        self.assertEqual(data.output["docker_image"], "resolwe/base:ubuntu-20.04")
+        self.assertEqual(data.output["docker_image"], "resolwe/base:ubuntu-20.04dev")
         self.assertEqual(data.name, "Foo: bar")
 
         entity = Entity.objects.get(data=data)