+++ docs/CHANGELOG.rst
@@ -30,6 +30,7 @@ Added
 -----
 - Make automatic removal of Docker containers configurable
 - Terminate processing immediately when data object is deleted
+- Make default processing image configurable
 
 
 ===================